# Introduce aggregate cache in database layer