--- a/daemon/src/auto_rollover.rs
+++ b/daemon/src/auto_rollover.rs
@@ -3,10 +3,9 @@ use crate::db;
 use crate::oracle;
 use crate::process_manager;
 use crate::rollover_taker;
-use crate::try_continue;
-use anyhow::Context;
 use anyhow::Result;
 use async_trait::async_trait;
+use futures::StreamExt;
 use maia::secp256k1_zkp::schnorrsig;
 use model::OrderId;
 use std::time::Duration;
@@ -100,26 +99,34 @@ where
         &mut self,
         ctx: &mut xtra::Context<Actor<O>>,
     ) -> Result<(), anyhow::Error> {
-        let mut conn = self.db.acquire().await?;
-        let cfd_ids = db::load_open_cfd_ids(&mut conn).await?;
         let this = ctx
             .address()
             .expect("actor to be able to give address to itself");
-        for id in cfd_ids {
-            try_continue!(async {
-                let cfd = db::load_cfd::<model::Cfd>(id, &mut conn, ()).await?;
 
-                match cfd.can_auto_rollover_taker(OffsetDateTime::now_utc()) {
-                    Ok(()) => this.send_async_safe(Rollover(id)).await?,
-                    Err(reason) => {
-                        tracing::trace!(order_id = %id, %reason, "CFD is not eligible for auto-rollover");
-                    }
+        let mut conn = self.db.acquire().await?;
+        let mut stream = db::load_all_open_cfds::<model::Cfd>(&mut conn, ());
+
+        while let Some(cfd) = stream.next().await {
+            let cfd = match cfd {
+                Ok(cfd) => cfd,
+                Err(e) => {
+                    tracing::warn!("Failed to load CFD from database: {e:#}");
+                    continue;
                 }
-                anyhow::Ok(())
+            };
+            let id = cfd.id();
+
+            match cfd.can_auto_rollover_taker(OffsetDateTime::now_utc()) {
+                Ok(()) => {
+                    let _ = this.send_async_safe(Rollover(id)).await; // If we disconnect, we don't
+                                                                      // care.
+                }
+                Err(reason) => {
+                    tracing::trace!(order_id = %id, %reason, "CFD is not eligible for auto-rollover");
+                }
             }
-            .await
-            .context("Cannot roll over"));
         }
+
         Ok(())
     }
 }