--- a/daemon/src/collab_settlement_taker.rs
+++ b/daemon/src/collab_settlement_taker.rs
@@ -1,5 +1,6 @@
 use crate::command;
 use crate::connection;
+use crate::db;
 use crate::process_manager;
 use crate::wire;
 use anyhow::anyhow;
@@ -26,7 +27,7 @@ pub struct Actor {
     n_payouts: usize,
     connection: xtra::Address<connection::Actor>,
     executor: command::Executor,
-    db: sqlx::SqlitePool,
+    db: db::Connection,
     tasks: Tasks,
     maker_replied: bool,
 }
@@ -38,7 +39,7 @@ impl Actor {
         n_payouts: usize,
         connection: xtra::Address<connection::Actor>,
         process_manager: xtra::Address<process_manager::Actor>,
-        db: sqlx::SqlitePool,
+        db: db::Connection,
     ) -> Self {
         Self {
             proposal: None,
@@ -82,9 +83,7 @@ impl Actor {
 
         tracing::info!(%order_id, "Settlement proposal got accepted");
 
-        let mut conn = self.db.acquire().await?;
-
-        let cfd = crate::db::load_cfd::<model::Cfd>(&mut conn, order_id, ()).await?;
+        let cfd = self.db.load_cfd::<model::Cfd>(order_id, ()).await?;
 
         // TODO: This should happen within a dedicated state machine returned from
         // start_collaborative_settlement