--- a/daemon/src/lib.rs
+++ b/daemon/src/lib.rs
@@ -52,7 +52,6 @@ pub mod setup_maker;
 pub mod setup_taker;
 pub mod taker_cfd;
 mod transaction_ext;
-pub mod try_continue;
 pub mod version;
 pub mod wallet;
 pub mod wire;