See commit messages for details. This solution does have _some_ trade-offs but should scale a lot better than the caches that we previously had. Hopefully, this will improve the overall responsiveness of our application. Some of the database calls on mainnet take shockingly long which may be the cause for some of the aborted protocols?

Flamegraphing didn't show much difference but that is somewhat expected. Flamegraphs only show how the given CPI time is spent but only relative to how long things actually take. We still need to deserialize the events at some point, even with this patch set.

Where we _can_ see a difference is in the metrics:

With a 5GB database:

- On master, `projection::Initialize` takes 275 seconds
- On master, 12 `NewAttestationFetched` messages took 593 seconds on startup 
- With this patch-set, `projection::Initialize` went down to 142 seconds
- With this patch-set, 56(!) `NewAttestationFetched` messages took 88 seconds

When testing on master, I thought the app was fully initialized after 12 `NewAttestationFetched` messages but that didn't actually seem to be true. In any case, we can see a massive performance boost here!