@thomaseizinger 2022-03-30
Oh god. I think the tests are failing because they poison each others cache ...

I guess I need to actually associate the cache with the DB connection. Thought I could avoid that -.-