@thomaseizinger 2022-03-31
Force-push only contains @luckysori's feedback: https://github.com/itchysats/itchysats/compare/216f619786a6d966afd247176ac9b369c87676bf..620efb2cb1bce03db2a3650bcee2ecb03a05ed0c